add loop logic to fill whole screen white or black
there is odd behavior with the emulator when trying to fill the whole
screen (using writing to all 8192 words on the memory map) on no
animation mode. however, seems to work if you step through the program
-- however going through all 8192 x 2 iterations is more tedious than
i'd like. testing with smaller values (e.g. 3000) does work, and half
fills the screen as expected.